+++ lib/grape/validations/validators/exactly_one_of.rb
@@ -1,11 +1,11 @@
+require 'grape/validations/validators/multiple_params_base'
+
 module Grape
   module Validations
-    require 'grape/validations/validators/multiple_params_base'
     class ExactlyOneOfValidator < MultipleParamsBase
       def validate_params!(params)
-        if keys_in_common(params).length != 1
-          raise Grape::Exceptions::Validation, params: all_keys, message: message(:exactly_one)
-        end
+        return if keys_in_common(params).length == 1
+        raise Grape::Exceptions::Validation, params: all_keys, message: message(:exactly_one)
       end
     end
   end